fix: error delete nonexitent record

+++ api-template/handler/client.go
@@ -13,11 +13,10 @@ func (h *Handler) GetAllClient(c echo.Context) error {
 
 	classes, err := h.repo.Client.GetAllClients(h.store)
 
-	c.JSON(http.StatusOK, classes)
 	if err != nil {
 		return util.HandleError(c, errors.ErrInternalServerError)
 	}
-	return nil
+	return c.JSON(http.StatusOK, classes)
 }
 
 func (h *Handler) AddClient(c echo.Context) error {
@@ -40,23 +39,26 @@ func (h *Handler) AddClient(c echo.Context) error {
 	if err != nil {
 		return util.HandleError(c, errors.ErrInternalServerError)
 	}
-	c.JSON(http.StatusOK, client)
 
-	return nil
+	return c.JSON(http.StatusOK, client)
 }
 
 func (h *Handler) DeleteClient(c echo.Context) error {
 	id := c.Param("id")
+
+	//check user exist
+	user, _ := h.repo.Client.GetByID(h.store, id)
+
+	if user.Username == "" {
+		return util.HandleError(c, errors.ErrClientDoesNotExits)
+	}
 
 	_, err := h.repo.Client.Delete(h.store, id)
 
 	if err != nil {
 		return util.HandleError(c, errors.ErrInternalServerError)
 	}
-	c.JSON(http.StatusOK, "Delete successfully the clinent with id"+id)
-
-	return nil
-
+	return c.JSON(http.StatusOK, "Delete successfully the clinent with id"+id)
 }
 
 func (h *Handler) UpdateClientPasswordByID(c echo.Context) error {
@@ -67,9 +69,7 @@ func (h *Handler) UpdateClientPasswordByID(c echo.Context) error {
 	if err != nil {
 		return util.HandleError(c, errors.ErrInternalServerError)
 	}
-	c.JSON(http.StatusOK, client)
-
-	return nil
+	return c.JSON(http.StatusOK, client)
 }
 
 func (h *Handler) GetRole(c echo.Context, username string, password string) (string, error) {
@@ -89,8 +89,5 @@ func (h *Handler) Login(c echo.Context) error {
 	if err != nil {
 		return util.HandleError(c, errors.ErrAccountNotfound)
 	}
-	c.JSON(http.StatusOK, client)
-
-	return nil
-
+	return c.JSON(http.StatusOK, client)
 }
